kadai-tasklist-login

+++ app/Http/Controllers/TasksController.php
@@ -50,9 +50,10 @@ class TasksController extends Controller
        $task = new Task;
        $task->status = $request->status;
         $task->content = $request->content;
+        $task->user_id=$request->user()->id;
         $task->save();
 
-        return redirect('/');
+        return redirect('/tasks');
     }
 
     /**
@@ -64,10 +65,15 @@ class TasksController extends Controller
     public function show($id)
     {
         $task = Task::find($id);
-
+      
+        if (\Auth::id() === $task->user_id) {
+        
         return view('tasks.show', [
             'task' => $task,
-        ]);
+        ]);}else{
+           
+             return redirect('/tasks');
+        }
     }
 
     /**
@@ -79,10 +85,17 @@ class TasksController extends Controller
     public function edit($id)
     {
          $task = Task::find($id);
-
+        
+        if (\Auth::id() === $task->user_id) {
+            
         return view('tasks.edit', [
             'task' => $task,
-        ]);
+        ]);}else{
+            
+             return redirect('/tasks');
+        }
+        
+       
     }
 
     /**
@@ -103,7 +116,7 @@ class TasksController extends Controller
         $task->content = $request->content;
         $task->save();
 
-        return redirect('/');
+        return redirect('/tasks');
     }
 
     /**
@@ -115,8 +128,10 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::find($id);
-        $task->delete();
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
 
-        return redirect('/');
+        return redirect('/tasks');
     }
 }
